+++ app/src/main/java/com/flydrop2p/flydrop2p/domain/repository/OwnAccountRepository.kt
@@ -6,6 +6,6 @@ import kotlinx.coroutines.flow.Flow
 interface OwnAccountRepository {
     fun getAccountAsFlow(): Flow<Account>
     suspend fun getAccount(): Account
-    suspend fun setAccountId(accountId: Int)
+    suspend fun setAccountId(accountId: Long)
     suspend fun setProfileUpdate(profileUpdate: Long)
 }
